+++ domain/src/main/java/ru/aleksandrorlov/domain/interactor/GetYandexPictureDetails.java
@@ -30,13 +30,13 @@ public class GetYandexPictureDetails extends UseCase<YandexPicture, GetYandexPic
 
     public static final class Params {
 
-        private final int yandexPictureId;
+        private final long yandexPictureId;
 
-        private Params(int yandexPictureId) {
+        private Params(long yandexPictureId) {
             this.yandexPictureId = yandexPictureId;
         }
 
-        public static Params forYandexPicture(int yandexPictureId) {
+        public static Params forYandexPicture(long yandexPictureId) {
             return new Params(yandexPictureId);
         }
     }
